requirements.txt
discord.py
discord-py-slash-command
python-dotenv
pytz
requests
openai
python-dotenv
dotenv